remove unused files
